Correction in event component
