--- src/app/events/events-index/events-index.component.ts
@@ -1,21 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-events-index',
-  templateUrl: './events-index.component.html',
-  styleUrls: ['./events-index.component.css']
-})
-export class EventsIndexComponent implements OnInit {
-
-  constructor() { }
-
-  xs: string[] = ["aa", "bb", "cc", "dd", "ee", "ff", "gg", "hh"];
-
-  ngOnInit() {
-  }
-
-
-  
-
-
-}
